+++ src/metrology/shell.clj
@@ -3,11 +3,11 @@
 (require '[clojure.java.shell :refer [sh]])
 
 (pprint (gs2000 1
-                ;"CH4"
-                496
-                (list 9 25 45)
+                "NH3"
+                3000
+                (list 40 300 550)
                 #_(map #(ch/ppm->mg "H2S" %1)
-                     (list 5 9.5 14 25 47))))
+                     (list 5.84 29))))
 
 ((gs/calculator
   (gs/passports 1))
@@ -19,41 +19,45 @@
 ;; #chemistry
 (* (ch/coefficient "C6H14") 900)
 
-(ch/ppm->mg "H2S" 95)
+(ch/ppm->mg "H2S" 90)
 
-(ch/mg->ppm "CH4" 1500)
+(ch/mg->ppm "C6H14" 3500)
 
 (map #(/ (* %1 32.07) 62.14)
      '(4.9 7.8 40 70))
 
+(map #(/ %1 4.4 0.01)
+     '(2.18 4.22))
+
 ;; #report#methodology
-(methodology (list 209))
+(methodology (list 61))
 (sh "vivaldi" (str midb-path "methodology.html"))
 
 ;; #find#methodology
 (methodology
   (map (fn [m]
            (:id m))
-       (find-methodology "Сигнал")))
+       (find-methodology "021")))
 (sh "vivaldi" (str midb-path "methodology.html"))
 
 ;; #find#record
 (gen-report
   (find-verification
-    "v.id >= 3230"))
+    "v.id >= 3403"))
 (sh "vivaldi" (str midb-path "report.html"))
 
 ;; #report#find#mi
 (gen-report
   (find-verification
-    "lower(v.mi_type) like '%ФП22%'
-     --and components like '%ПФД%'
-     --and channels = 4
-     --and components like '%NH3%'
-     --and met.registry_number like '%-07%'
+    "lower(v.mi_type) like '%ХОББИТ-%'
+     --and channels = 1
+     --and components like '%SO%'
+     --and components like '%O2%'
+     --and v.comment not like 'Леонтьев'
+     --and met.registry_number like '%-86%'
      --v.protocol_number = 2124
-     --v.serial_number like '%412-1014930%'
-     --count like '%29768%'"))
+     --and v.serial_number like '%219-025786%'
+     --and count like '%3007952%'"))
 (sh "vivaldi" (str midb-path "report.html"))
 
 ;; #report
@@ -61,22 +65,22 @@
 (sh "vivaldi" (str midb-path "report.html"))
 
 ;; #report#protocols
-(gen-protocols "id >= 3236")
+(gen-protocols "id >= 3403")
 (sh "vivaldi" (str midb-path "protocol.html"))
 
 ;; #gen#measurements#values
-(gen-values! "id >= 3236")
+(gen-values! "id >= 3403")
 
 ;; #find#counteragents
-(counteragents "Ю%У%Г%ЕРГО")
+(counteragents "ГАЗПРОМТРАН")
 (sh "vivaldi" (str midb-path "counteragents.html"))
 
 ;; #copy#record
-(map (fn [s] (copy-record! 2980))
+(map (fn [s] (copy-record! 3411))
      (range 1))
 
 (let [nums (map (fn [n] (str "" n))
-                (list 2010079)
+                (list 1903063)
                 #_(range 18))
       start-id (next-id)
       start-protocol-number (next-protocol-number)]
@@ -87,14 +91,14 @@
            :verification
            (hash-map
              ;:methodology_id 375
-             ;:mi_type "ОКА-92МТ-O₂-CH₄-CO-H₂S-И13"
-             ;:components "CH₄ (0 - 2,2) % об.; O₂ (0 - 30) % об.; CO (0 - 200) мг/м³; NH3 (0 - 70) мг/м³"
-             ;:channels 3
-             :count "9/3007851"
-             :counteragent 8
-             :conditions 1100
-             :manufacture_year 2020
-             ;:comment "Леонтьев"
+             :mi_type "ОКА-М-CH₄-И13"
+             ;:components "CH₄ (метан); O₂ (кислород); SO₂ (диоксид серы); H₂S (сероводород)"
+             :channels 1
+             :count "9/3007902"
+             :counteragent 14228
+             :conditions 1110
+             :manufacture_year 2019
+             :comment "Леонтьев"
              ;:comment 11
              ;:comment "ГИС блок 2"
              ;:upload 1
@@ -102,10 +106,10 @@
              ;:scope
              ;:engineer 3514
              ;:sw_name "MXTF_11_000"
-             ;:sw_version "не ниже 12" 
+             ;:sw_version "не ниже v01.00" 
              ;:sw_checksum "8BFD"
              ;:sw_algorithm "CRC 16"
-             ;:sw_version_real "50D"
+             ;:sw_version_real "5.15"
              :serial_number n
              :protocol_number (+ start-protocol-number i)
              :protocol nil
@@ -116,7 +120,7 @@
        (range (count nums))))
 
 ;; #delete#record
-(delete-record! 3157)
+(delete-record! 3411)
 
 ;; Удалить записи с id >=
 ;; #delete#record
@@ -134,15 +138,25 @@
 
 ;; #set#gso
 (set-v-gso!
-  3238
-  (list 377)
-  #_(map (fn [m]
+  3417
+  #_(list 382)
+  (map (fn [m]
            (:id m))
-       (check-gso (list "14638-23" "14636-23" "14634-23" "02464-23" "00810-23" "16782-23")
+       (check-gso (list "14632-23")
                   "pass_number")))
 
+;; #update#gso
+(jdbc/update!
+  midb
+  :v_gso
+  {
+   :gso_id 377
+  }
+  ["v_id = ? 
+    and gso_id = ?" 3349 332])
+
 ;; #copy#gso
-(copy-v-gso! 2574 2579)
+(copy-v-gso! 3255 3256)
 
 ;; #delete#gso
 (delete-v-gso! 2343)
@@ -151,23 +165,23 @@
             (remove #{258 334} (get-v-gso 2386)))
 
 ;; #conditions
-(conditions "2023-11-29")
+(conditions "2023-12-13")
 (sh "vivaldi" (str midb-path "conditions.html"))
 
 ;; #add#conditions
-(insert-conditions! {:date "2023-12-01"
-                     :temperature 22.4
-                     :humidity 51.4
-                     :pressure 100.7
+(insert-conditions! {:date "2023-12-14"
+                     :temperature 22.2
+                     :humidity 52.3
+                     :pressure 103.33
                      :voltage 220.0
                      :frequency 50
-                     ;:other "расход ГС (0,1 - 0,3) л/мин."
+                     :other "расход ГС 0,4 (0,4 ± 0,05) дм³/мин."
                      ;:location "с. Ивановка"
                      ;:comment ""
                      })
 
 ;; #set#refs
-(set-v-refs! 3238
+(set-v-refs! 3413
              (list 2663 2820))
 
 ;; #copy#refs
@@ -182,7 +196,7 @@
   {:v_id 2343 :ref_id 2762})
 
 ;; #set#opt-refs
-(set-v-opt-refs! 3237
+(set-v-opt-refs! 3307
                  (list 2643 2831 2762 2756))
 
 ;; #copy#opt-refs
@@ -201,17 +215,17 @@
    })
 
 ;; #set#operations
-(set-v-operations! 3238
-                   (list 1752 1753 1754))
+(set-v-operations! 3307
+                   (list 258 532 806 1171))
 
 ;; #copy#operations
 (copy-v-operations! 3187 3210)
 
 ;; #unusability#update#operations
 (unusability
-  3230
-  766
-  "ошибка заряда, не включается")
+  3410
+  499
+  "негерметичен")
 
 ;Проверить ГСО в записи.
 (pprint (check-gso (map (fn [x] (:gso_id x))
@@ -229,16 +243,16 @@
   :verification
   (hash-map
      :engineer 3514
-     :count "9/3007851"
-     :counteragent 8 
-     :conditions 1100
+     :count "9/030057"
+     :counteragent 168
+     :conditions 1105
      :verification_type 1
      :protocol_number (next-protocol-number) 
      ;:protocol_number 3182
-     :mi_type "Сигнал-02М"
-     :methodology_id 27
-     :serial_number "6976-01"
-     :manufacture_year 2010
+     :mi_type "Лидер, мод. Лидер-021"
+     :methodology_id 332
+     :serial_number "D919011009"
+     :manufacture_year 2019
      :channels 1
      :area "05"
      :interval 12
@@ -251,7 +265,7 @@
      ;:sw_version_real "V4.0"
      ;:voltage 24
      ;:upload 
-     ;:comment "Леонтьев"
+     :comment "Леонтьев"
      ))
 
 (jdbc/update!
@@ -349,8 +363,9 @@
 
 ;; #add#measurements
 (add-measurements
-  3238
-  (list [1508 0] [1508 20] [1508 50] [1509 nil]
+  3417
+  (list [638 0 0.15] [639 0.25 0.45 0.25] [638 0.15 0]
+        [639 0.45] [641 nil]
         ))
 
 ;; #update#measurements
@@ -370,7 +385,7 @@
      (range 2671 2689))
 
 ;; #delete#measurements
-(delete-measurements! 3215)
+(delete-measurements! 3417)
 
 (jdbc/delete!
   midb
@@ -380,44 +395,44 @@
 
 ;; #add#metrology#channel
 (ins-channel!
-  {:methodology_id 27
+  {:methodology_id 68
    :channel nil
-   :component "CH4"
+   :component "SO2"
    :range_from 0
-   :range_to 50
-   :units "% НКПР"
-   :low_unit 5
+   :range_to 100
+   :units "мг/м³"
+   :low_unit 1
    :view_range_from 0
-   :view_range_to 50
-   ;:comment "Sensepoint; диапазон показаний условно!"
+   :view_range_to 100
+   :comment "диапазон показаний условно!"
    }
   (list {:r_from 0
-         :r_to 50
-         :value 5
+         :r_to 10
+         :value 25
          :fraction nil
-         :type_id 0
+         :type_id 2
          :units nil
-         :operation_id 1753 
+         :operation_id 589
          :comment nil}
-        #_{:r_from 10
+        {:r_from 10
          :r_to 100
-         :value 20
+         :value 25
          :fraction nil
          :type_id 1
          :units nil
-         :operation_id 1106 
+         :operation_id 589
          ;:comment "(15 - 30) % об."
          }
-        #_{:value 0.5
+        {:value 0.5
          :type_id 5
          :units ""
-         :operation_id 1237}
-        #_{;:r_from 0
+         :operation_id 850}
+        {;:r_from 0
          ;:r_to 10
-         :value 40
+         :value 120
          :type_id 6
          :units "с"
-         :operation_id 1333}
+         :operation_id 1048}
         #_{
          :r_from 0
          :r_to 71 
@@ -431,10 +446,10 @@
          :operation_id 878
          ;:comment "порог 11 % НКПР"
         }
-        {:value 10
+        #_{:value 10
          :type_id 7
          :units "с"
-         :operation_id 1754}
+         :operation_id 1757}
         #_{:value 30
          :type_id 16
          :units "с"
@@ -445,10 +460,10 @@
 (jdbc/insert!
   midb
   :counteragents
-  {:name "ОБЩЕСТВО С ОГРАНИЧЕННОЙ ОТВЕТСТВЕННОСТЬЮ «ЭНКИСТРОЙМОНТАЖ»"
-   :short_name "ООО «ЭСМ»"
-   :address "460021, Оренбургская область, город Оренбург, Красногорская ул., д. 72/1"
-   :inn "5610227756"})
+  {:name "ФИЛИАЛ АО «ГАЗПРОМ ГАЗОРАСПРЕДЕЛЕНИЕ ОРЕНБУРГ» В Г. МЕДНОГОРСКЕ (МЕДНОГОРСКМЕЖРАЙГАЗ)"
+   :short_name "МЕДНОГОРСКМЕЖРАЙГАЗ"
+   :address "462250, Оренбургская область, город Медногорск, ул. Кирова, д. 10"
+   :inn "5610010369"})
 
 ;; #update#counteragents
 (jdbc/update!
@@ -457,9 +472,9 @@
   {
    ;:name "ООО АТЦ «Жигули-Оренбург»"
    ;:short_name "ООО АТЦ «Жигули-Оренбург»"
-   :address "460044, Оренбургская обл, Оренбург г, Терешковой ул, дом № 154/2"
+   :address "462781, Оренбургская обл, Комаровский п, Комарова ул, дом 3, корпус А"
    }
-  ["id = ?" 41551])
+  ["id = ?" 14228])
 
 ;; #cars
 ;; #insert#order#auto
@@ -468,12 +483,12 @@
     auto
     :travel_order
     {:auto_id 2
-     :count "9/"
-     :date_departure "2023-11-28T09:00"
-     :date_arrive "2023-11-28T14:30"
-     :odometr_departure 136746
-     :fuel_departure 2.21
-     :odometr_arrive 136975
+     :count "9/3007981"
+     :date_departure "2023-12-08T10:00"
+     :date_arrive "2023-12-08T13:30"
+     :odometr_departure 137964
+     :fuel_departure 12.97
+     :odometr_arrive 137984
      :fuel_add 30})
   (pprint
     (jdbc/query
@@ -596,13 +611,15 @@
 
 (require '[clojure.repl :refer :all])
 
-(find-doc "assoc")
+(find-doc "air")
 
 (doc get-in)
 
-(dir clojure.math)
+(dir metrology.lib.metrology)
 
-(doc string/split)
+(doc metr/air-v->vnc)
+
+(metr/air-v->vnc 100 23.4 103.87)
 
 (dir clojure.core)